requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
flask>=2.3.0
flask-cors>=4.0.0
redis>=5.0.0
diskcache>=5.0.0
pandas>=2.0.0
python-json-logger>=2.0.0
typer>=0.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0